projects: cn0565: restruct to EXAMPLE build
Move the CN0565 EIT app to the EXAMPLE-based flow, splitting the shared
main.c into per-platform dispatchers (stm32, aducm3029) with an eit and
an eit_iio example over common data, mirroring eval-adxrs290-pmdz. Update
the STM32 CMake/Kconfig/.conf wiring for the eit/eit_iio variants.
Behavior unchanged.
