+++ projects/cn0565/builds.json
@@ -1,10 +1,10 @@
 {
   "aducm3029": {
     "fw_eit_ad5940": {
-      "flags": ""
+      "flags": "EXAMPLE=eit"
     },
     "fw_eit_ad5940_iio": {
-      "flags": "IIOD=y"
+      "flags": "EXAMPLE=eit_iio"
     }
   }
 }
