+++ projects/cn0565/eit.conf
@@ -8,3 +8,4 @@ CONFIG_DMA=y
 CONFIG_TIMER=n
 CONFIG_AFE=y
 CONFIG_AFE_AD5940=y
+CONFIG_CN0565_EIT_EXAMPLE=y
