+++ projects/cn0565/eit_iio.conf
@@ -11,3 +11,4 @@ CONFIG_AFE=y
 CONFIG_AFE_IIO_AD5940=y
 CONFIG_SWITCH=y
 CONFIG_SWITCH_IIO_ADG2128=y
+CONFIG_CN0565_EIT_IIO_EXAMPLE=y
