+++ projects/cn0565/src/examples/eit/example.mk
@@ -1,0 +1,1 @@
+CFLAGS += -DENABLE_UART_STDIO
