+++ projects/cn0565/src/examples/eit/mux_board.c
