+++ projects/cn0565/src/platform/stm32/parameters.c
@@ -1,0 +1,57 @@
+/***************************************************************************//**
+ *   @file   parameters.c
+ *   @brief  Definition of STM32 platform data used by cn0565 project.
+ *   @author Kister Genesis Jimenez (kister.jimenez@analog.com)
+********************************************************************************
+ * Copyright 2023(c) Analog Devices, Inc.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of Analog Devices, Inc. nor the names of its
+ *    contributors may be used to endorse or promote products derived from this
+ *    software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY ANALOG DEVICES, INC. “AS IS” AND ANY EXPRESS OR
+ * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
+ * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
+ * EVENT SHALL ANALOG DEVICES, INC. BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA,
+ * OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
+ * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*******************************************************************************/
+
+#include "parameters.h"
+
+struct stm32_uart_init_param cn0565_uart_extra_ip = {
+	.huart = &huart5,
+};
+
+struct stm32_spi_init_param cn0565_spi_extra_ip = {
+	.chip_select_port = SPI_CS_PORT,
+	.get_input_clock = HAL_RCC_GetPCLK1Freq,
+};
+
+struct stm32_gpio_init_param cn0565_reset_gpio_extra_ip = {
+	.mode = GPIO_MODE_OUTPUT_PP,
+	.speed = GPIO_SPEED_FREQ_VERY_HIGH,
+};
+
+struct stm32_gpio_init_param cn0565_gp0_gpio_extra_ip = {
+	.mode = GPIO_MODE_INPUT,
+	.speed = GPIO_SPEED_FREQ_VERY_HIGH,
+};
+
+struct stm32_gpio_irq_init_param cn0565_gpio_irq_extra_ip = {
+	.port_nb = 6, /* port G */
+};
