+++ projects/cn0565/src/platform/stm32/platform_src.mk
@@ -1,0 +1,20 @@
+SRCS += $(PLATFORM_DRIVERS)/stm32_delay.c \
+	$(PLATFORM_DRIVERS)/stm32_gpio.c \
+	$(PLATFORM_DRIVERS)/stm32_spi.c \
+	$(PLATFORM_DRIVERS)/stm32_dma.c  \
+	$(PLATFORM_DRIVERS)/stm32_i2c.c \
+	$(PLATFORM_DRIVERS)/stm32_irq.c \
+	$(PLATFORM_DRIVERS)/stm32_gpio_irq.c \
+	$(PLATFORM_DRIVERS)/stm32_uart.c \
+	$(PLATFORM_DRIVERS)/stm32_uart_stdio.c
+
+INCS +=	$(INCLUDE)/no_os_delay.h \
+	$(PLATFORM_DRIVERS)/stm32_uart_stdio.h \
+	$(PLATFORM_DRIVERS)/stm32_uart.h \
+	$(PLATFORM_DRIVERS)/stm32_irq.h \
+	$(PLATFORM_DRIVERS)/stm32_gpio_irq.h \
+	$(PLATFORM_DRIVERS)/stm32_spi.h \
+	$(PLATFORM_DRIVERS)/stm32_dma.h  \
+	$(PLATFORM_DRIVERS)/stm32_i2c.h \
+	$(PLATFORM_DRIVERS)/stm32_gpio.h \
+	$(PLATFORM_DRIVERS)/stm32_hal.h
